Generate declarative config for app and appproj
