+++ docs/operator-manual/server-commands/argocd-util_declarative_config.md
@@ -1,0 +1,24 @@
+## argocd-util config
+
+Generate declarative config
+
+### Synopsis
+
+Generate declarative config
+
+```
+argocd-util config [flags]
+argocd-util config [command]
+```
+
+### Options
+
+```
+  -h, --help   help for config
+```
+
+### SEE ALSO
+
+* [argocd-util](argocd-util.md)	 - argocd-util tools used by Argo CD
+* [argocd-util config app](argocd-util_declarative_config_app.md) - Check declarative config for an application
+* [argocd-util config proj](argocd-util_declarative_config_proj.md) - Check declarative config for a project
